project/backend/requirements.txt
## The following requirements were added by pip freeze:
alembic==1.13.3
SQLAlchemy==2.0.35
tornado==6.4.1
pyjwt==2.9.0